requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
numpy>=1.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
